+++ src/main/java/com/rsd_movieshop/service/UserService.java
@@ -1,11 +1,16 @@
 package com.rsd_movieshop.service;
 
+import com.rsd_movieshop.dto.UserDto;
 import com.rsd_movieshop.model.User;
 import com.rsd_movieshop.repository.UserRepo;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class UserService {
@@ -14,24 +19,74 @@ public class UserService {
 
 
     public UserService(UserRepo userRepo) {
-        this.userRepo = userRepo;
+		this.userRepo = userRepo;
+    }
+    
+
+    public ResponseEntity<User> findUserById(long id){
+    	if (id <= 0) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	}
+    	if (userRepo.getById(id) == null) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "The user with id: " + id + " doesn't exist!");
+    	}else {
+    		try {
+    			User user = userRepo.getById(id);
+    			return new ResponseEntity<User>(user, HttpStatus.OK);
+			} catch (Exception e) {
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e.getCause());
+			}
+    	}
     }
 
-    public Optional<User> findUserById(long id){
-        return userRepo.findById(id);
+    public ResponseEntity<List<User>> findAllUser(){
+    	try {
+    		List<User> users = userRepo.findAll();
+    		return new ResponseEntity<List<User>>(users, HttpStatus.OK);
+		} catch (Exception e) {
+			throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e.getCause());
+		}
     }
 
-    public List<User> findAllUser(){
-        return userRepo.findAll();
+    public ResponseEntity<User> saveUser(UserDto userDto){
+    	if (userDto.email == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	}
+    	if (userDto.familyName == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	}
+    	if (userDto.firstName == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	}
+    	if (userDto.userName == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	}
+    	if (userDto.password == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_ACCEPTABLE);
+    	} else {
+    		try {
+    			User user =  new User();
+    			user.setEmail(userDto.email);
+    			user.setFamilyName(userDto.familyName);
+    			user.setFirstName(userDto.firstName);
+    			user.setUserName(userDto.userName);
+    			BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
+    			String encodedPassword = passwordEncoder.encode(userDto.password);
+    			user.setPassword(encodedPassword);
+    			userRepo.save(user);
+				return new ResponseEntity<User>(user, HttpStatus.OK);
+			} catch (Exception e) {
+				throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage(), e.getCause());
+			}
+    	}
     }
 
-    public void saveUser(User user){
-    	
-        userRepo.save(user);
-    }
-
-    public void deleteUser(long id){
+    public ResponseEntity<String> deleteUser(long id){
+    	if (userRepo.getById(id) == null) {
+    		throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+    	}
         userRepo.deleteById(id);
+        return new ResponseEntity<String>("The user with the Id = " + id + " is deleted!", HttpStatus.OK);
     }
-
+    
 }
